- part 3

+++ code/astar.h
@@ -182,7 +182,7 @@ class AStarClass
 		 */
 		bool Find_Path_Hierarchical(Cell const & from, Cell const & to, MZoneType mzone, FootClass const * foot);
 		void Ban_Blocked_Subzone_Edges(FootClass const * foot);
-		bool Subzone_Edge_Banned(unsigned short subzone1, unsigned short subzone2, int subzone_level);
+		bool Subzone_Edge_Banned(int subzone1, int subzone2, int subzone_level); // was (unsigned short, unsigned short, int)
 		void Ban_Subzone_Edge(unsigned int subzone1, unsigned int subzone2, int subzone_level);
 		void Ban_Neighborhood_Subzone_Edges(unsigned int subzone, int subzone_level);
 
@@ -354,12 +354,12 @@ class AStarClass
 		 * per hierarchy level; the hierarchical search refuses to expand across
 		 * these when it retries after a regular pathfinding failure
 		 */
-		DynamicVectorClass<unsigned int> HierBannedEdges[SUBZONE_COUNT];
+		DynamicVectorClass<unsigned long long> HierBannedEdges[SUBZONE_COUNT]; // was DynamicVectorClass<unsigned int> -- see Ban_Subzone_Edge (astar.cpp)
 
 		/*
 		 * Ordered list of subzone IDs forming the hierarchical path per level
 		 */
-		unsigned short HierSubzonePath[SUBZONE_COUNT][500];
+		int HierSubzonePath[SUBZONE_COUNT][500]; // was `unsigned short` -- truncated once a level exceeds 65535 subzones
 
 		/*
 		 * Number of valid entries in each hierarchical subzone path
